+++ src/parser/inline/extension/index.test.ts
@@ -13,10 +13,9 @@ describe('Unit: parser/inline/extension/index', () => {
       assert.deepStrictEqual(inspect(parser('[#]]')), undefined);
       assert.deepStrictEqual(inspect(parser('[# ]')), undefined);
       assert.deepStrictEqual(inspect(parser('[# a]')), undefined);
-      assert.deepStrictEqual(inspect(parser('[#a ]')), undefined);
+      assert.deepStrictEqual(inspect(parser('[# a ]')), undefined);
       assert.deepStrictEqual(inspect(parser('[#\\ ]')), undefined);
       assert.deepStrictEqual(inspect(parser('[#\\ a]')), undefined);
-      assert.deepStrictEqual(inspect(parser('[#a\\ ]')), undefined);
       assert.deepStrictEqual(inspect(parser('[#\n]')), undefined);
       assert.deepStrictEqual(inspect(parser('[#\\\n]')), undefined);
       assert.deepStrictEqual(inspect(parser('[#\\]')), undefined);
@@ -25,7 +24,9 @@ describe('Unit: parser/inline/extension/index', () => {
 
     it('basic', () => {
       assert.deepStrictEqual(inspect(parser('[#a]')), [['<a href="#index:a" rel="noopener">a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a ]')), [['<a href="#index:a" rel="noopener">a </a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#a b]')), [['<a href="#index:a-b" rel="noopener">a b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a\\ ]')), [['<a href="#index:a" rel="noopener">a </a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#a\\ b]')), [['<a href="#index:a-b" rel="noopener">a b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#[]]')), [['<a href="#index:[]" rel="noopener">[]</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#\\]]')), [['<a href="#index:]" rel="noopener">]</a>'], '']);